refactor(xmlrpc): extract client tests to sibling file
